Refactor modifiers to use percentage values and add overload mechanic
Standardized modifier values to use integer percentages for consistency (e.g., HealOnHit, Reinforce, Team Heal). Added overload mechanic to Entity, allowing overloaded entities to deal damage to a random enemy. Updated dice and ability logic to support these changes, including AP cost reduction, improved temporary modifier handling, and enhanced debug logging. Also added a CameraManager to the Loop scene and improved dice side construction.

+++ Dicebound Tactics/Assets/Scripts/Dice/DiceModifiers/HealOnHitModifier.cs
@@ -5,7 +5,7 @@ using TacticsToolkit;
 [CreateAssetMenu(menuName = "DiceModifiers/HealOnHit")]
 public class HealOnHitModifier : DiceModifier
 {
-    public float percent = 15f;
+    public int percent = 15;
 
     public override void Apply(Entity user)
     {
